reachcheck_mvp/requirements.txt
jinja2
weasyprint
xhtml2pdf
fastapi
uvicorn
requests
python-dotenv
openai
beautifulsoup4
playwright
